refactor: consolidate model files into model_dalr.py
Remove the legacy src/model_dalr.py and rename src/model_modified.py
(the version actually used by train_mix.py) to src/model_dalr.py.
Update the import in train_mix.py accordingly.
